Refactored some more memory components and tri-state logic should now work as it is expected
You cannot set the same wire to more than one value distinct value (H/L), unless that other value is Z. Since Z doesn't carry any kind of value, it should have no effect, and only the actual values H/L will affect the value of the wire, if some pin is attempting to set the wire to that value. If there are multiple output pins are connected to the same wire (e.g. a bus), then if more than one of the outputs are attempting to drive the wire, the circuit will generate an error, and shine that wire in red.