Nodes now need to read an expected amount of bit of data, allows for easier error management.
ROMs and RAMs now have proper word addressing, so the WordSize property will allow the user to specify how many bytes of data that should be available or written at the DATA pin. Will give the data in a LSB fashion, so lower address bytes will be lower bits in the DATA word.